+++ R/yaml.R
@@ -78,3 +78,26 @@ yaml_value = function(x) {
   }
   x
 }
+
+#' Partition the YAML metadata and the body in a document
+#'
+#' Split a document into the YAML metadata (which starts with `---` in the
+#' beginning of the document) and the body. The YAML metadata will be parsed.
+#' @param x A character vector of the document content.
+#' @export
+#' @return A list of components `yaml` and `body`.
+#' @examples
+#' xfun::yaml_body(c('---', 'title: Hello', 'output: markdown::html_document', '---', '', 'Content.'))
+yaml_body = function(x) {
+  i = grep('^---\\s*$', x)
+  n = length(x)
+  res = if (n < 2 || length(i) < 2 || (i[1] > 1 && !all(is_blank(x[seq(i[1] - 1)])))) {
+    list(yaml = list(), body = x)
+  } else list(
+    yaml = x[i[1]:i[2]], body = c(rep('', i[2]), tail(x, n - i[2]))
+  )
+  if ((n <- length(res$yaml)) >= 3) {
+    res$yaml = yaml_load(res$yaml[-c(1, n)])
+  }
+  res
+}