+++ NEWS.md
@@ -6,6 +6,8 @@
 
 - Added a function `yaml_load()` to read YAML data when the **yaml** package is not available. It only supports a limited number of data types and is supposed to be used as a fallback method. See the help page `?xfun::yaml_load` for details.
 
+- Added a function `yaml_body()` to split a document into YAML metadata and the body.
+
 - `is_arm64()` also supports Linux now (thanks, @eitsupi, #74).
 
 - `is_blank()` returns a logical vector of the same length as the input vector now, indicating if each element of the input is blank. Previously it returns a logical scalar indicating whether *all* elements are blank.